+++ source/psi3d/snippetsgl.h
@@ -55,8 +55,8 @@ struct cPrimitives {
                 glBegin(GL_TRIANGLE_STRIP);
                 loopi(steps+1) {
 
-                    float inv0 = 1.0f / (float) sqrt(a*a+b0*b0+1.0f);
-                    float inv1 = 1.0f / (float) sqrt(a*a+b1*b1+1.0f);
+                    float inv0 = 1.0f / (float) sqrtf(a*a+b0*b0+1.0f);
+                    float inv1 = 1.0f / (float) sqrtf(a*a+b1*b1+1.0f);
                     float v0[] = { a*inv0, b0*inv0, dir*inv0 };
                     float v1[] = { a*inv1, b1*inv1, dir*inv1 };
 
@@ -92,8 +92,8 @@ struct cPrimitives {
                 glBegin(GL_TRIANGLE_STRIP);
                 loopi(steps+1) {
 
-                    float inv0 = 1.0f / (float) sqrt(a*a+b0*b0+1.0f);
-                    float inv1 = 1.0f / (float) sqrt(a*a+b1*b1+1.0f);
+                    float inv0 = 1.0f / (float) sqrtf(a*a+b0*b0+1.0f);
+                    float inv1 = 1.0f / (float) sqrtf(a*a+b1*b1+1.0f);
                     float v0[] = { a*inv0, dir*inv0, b0*inv0 };
                     float v1[] = { a*inv1, dir*inv1, b1*inv1 };
 
@@ -129,8 +129,8 @@ struct cPrimitives {
                 loopi(steps+1) {
 
                     //a = fma(a,1,0);
-                    float inv0 = 1.0f / (float) sqrt(a*a+b0*b0+1.0f);
-                    float inv1 = 1.0f / (float) sqrt(a*a+b1*b1+1.0f);
+                    float inv0 = 1.0f / (float) sqrtf(a*a+b0*b0+1.0f);
+                    float inv1 = 1.0f / (float) sqrtf(a*a+b1*b1+1.0f);
                     float v0[] = { dir*inv0, a*inv0, b0*inv0 };
                     float v1[] = { dir*inv1, a*inv1, b1*inv1 };
